# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: Хроники Восьми Земель API
  version: 0.10.0
  description: Конструктор новых персонажей. Готовые герои книги остаются в content/.
...
servers:
  - url: /
paths:
...
  /drafts/{draftId}:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    get:
      operationId: getDraft
      summary: Восстановить свой черновик после входа
      description: Авторизация ещё не подключена. Маршрут всегда возвращает 403, в том числе со старым токеном черновика. Доступ и контракт записи будут включены после проверки личности пользователя и владельца анкеты.
      responses:
        '403': {$ref: '#/components/responses/DraftsRequireAuthentication'}
    patch:
      operationId: patchDraft
      summary: Сохранить раздел своего черновика после входа
      description: Авторизация ещё не подключена. Маршрут всегда возвращает 403, в том числе со старым токеном черновика. Доступ и контракт записи будут включены после проверки личности пользователя и владельца анкеты.
      responses:
        '403': {$ref: '#/components/responses/DraftsRequireAuthentication'}
...
  /drafts/{draftId}/validate:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    post:
      operationId: validateDraft
      summary: Проверить свой черновик
      description: Авторизация ещё не подключена. Маршрут всегда возвращает 403, в том числе со старым токеном черновика. Доступ и контракт записи будут включены после проверки личности пользователя и владельца анкеты.
      responses:
        '403': {$ref: '#/components/responses/DraftsRequireAuthentication'}
  /drafts/{draftId}/complete:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    post:
      operationId: completeDraft
      summary: Завершить свой черновик
      description: Авторизация ещё не подключена. Маршрут всегда возвращает 403, в том числе со старым токеном черновика. Доступ и контракт записи будут включены после проверки личности пользователя и владельца анкеты.
      responses:
        '403': {$ref: '#/components/responses/DraftsRequireAuthentication'}
...
  /characters:
    post:
      operationId: createCharacter
      summary: Сохранить целую готовую анкету без черновика
      description: |
        Публичное создание по текущим правилам character-creation-v3. Создатель — NULL до авторизации.
        requestId — UUID одной попытки сохранения и ID будущей карточки. Повтор с той же
        нормализованной анкетой возвращает существующую карточку; другая анкета с тем же ID — 409.
        Сервер рассчитывает HP/AC и эффекты. Для изображений используйте multipart/form-data:
        character — JSON анкеты, portrait — выбранный портрет, icon:<abilityId> — до трёх иконок активных навыков.
        Файлы — бинарные PNG/JPEG: портрет до 1 МиБ и 1600 px, иконка до 128 КиБ и 512 px по длинной стороне.
        Лимит JSON — 256 КиБ, всего multipart — 2 МиБ. Изображения проверяются сервером,
        ссылки и метаданные создаются одной транзакцией вместе с героем. Черновики и редактирование остаются закрыты.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCharacter'}
          multipart/form-data:
            schema:
              type: object
              required: [character]
              maxProperties: 5
              additionalProperties: false
              properties:
                character: {$ref: '#/components/schemas/CreateCharacter'}
                portrait: {type: string, format: binary}
              patternProperties:
                '^icon:[a-z0-9]+(-[a-z0-9]+)*$': {type: string, format: binary}
            encoding:
              character: {contentType: application/json}
              portrait: {contentType: 'image/png, image/jpeg'}
            x-max-body-bytes: 2097152
      responses:
        '201':
          description: Персонаж, навыки и предметы сохранены одной транзакцией
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Character'}
        '200':
          description: Повтор ранее успешного сохранения, без новой записи
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Character'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409':
          description: requestId уже занят другой анкетой; код creation_conflict
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '413':
          description: Превышен лимит JSON или multipart; код request_too_large
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '422':
          description: Анкета не прошла проверку, ни одна строка не создана
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Validation'}
        '503': {$ref: '#/components/responses/Unavailable'}
        '500':
          description: Сбой сохранения; повтор с тем же requestId безопасен
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
    get:
      operationId: listCharacters
      summary: Список готовых персонажей конструктора
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: fresh
          in: query
          description: Обойти кэш процесса после сохранения, чтобы увидеть запись через любой экземпляр API.
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: Готовые персонажи, новые первыми
          headers:
            Server-Timing: {$ref: '#/components/headers/ServerTiming'}
            Cache-Control: {$ref: '#/components/headers/ReadCacheControl'}
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/CharacterSummary'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /characters/{characterId}:
    parameters:
      - name: characterId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getCharacter
      summary: Прочитать готового персонажа, навыки и предметы
      responses:
        '200':
          description: Готовый персонаж
          headers:
            Server-Timing: {$ref: '#/components/headers/ServerTiming'}
            Cache-Control: {$ref: '#/components/headers/ReadCacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Character'}
        '404': {$ref: '#/components/responses/NotFound'}
components:
  securitySchemes:
    AdvisorToken:
      type: http
      scheme: bearer
      description: Случайный секрет отдельной сессии советника, не авторизация игрока.
  headers:
    ServerTiming:
      description: Длительность обработки публичного чтения в процессе API (app;dur, миллисекунды) и результат серверного кэша (read_cache;desc, hit или miss). Не включает запуск контейнера и сетевой путь до него.
      schema: {type: string}
    ReadCacheControl:
      description: no-store. Браузерный HTTP-кэш не используется; актуальность публичных данных контролируют TanStack Query и короткий кэш процесса API.
      schema: {type: string, const: no-store}
  parameters:
...
    DraftId:
      name: draftId
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    CreateCharacter:
      type: object
      additionalProperties: false
      required: [requestId, rulesetId, formData]
      x-max-body-bytes: 262144
      properties:
        requestId: {type: string, format: uuid}
        rulesetId: {type: string, const: character-creation-v3}
        formData: {$ref: '#/components/schemas/CompleteCharacterForm'}
    CompleteCharacterForm:
      type: object
      additionalProperties: false
      required: [appearance, race, class, attributes, abilities, equipment]
      properties:
        appearance:
          type: object
          additionalProperties: false
          required: [displayName]
          properties:
            displayName: {type: string, minLength: 1, maxLength: 120, pattern: '\S'}
            pronouns: {type: string, maxLength: 4000}
            roleLabel: {type: string, maxLength: 4000}
            story: {type: string, maxLength: 4000}
            motivation: {type: string, maxLength: 4000}
            appearance: {type: string, maxLength: 4000}
            personality:
              type: array
              maxItems: 20
              items: {type: string, maxLength: 120}
        race:
          type: object
          additionalProperties: false
          required: [raceId]
          properties:
            raceId: {type: string}
        class:
          type: object
          additionalProperties: false
          required: [classId]
          properties:
            classId: {type: string}
        attributes:
          type: object
          additionalProperties: false
          required: [strength, dexterity, constitution, wisdom, intelligence, charisma]
          properties:
            strength: {type: integer, minimum: -4, maximum: 4}
            dexterity: {type: integer, minimum: -4, maximum: 4}
            constitution: {type: integer, minimum: -4, maximum: 4}
            wisdom: {type: integer, minimum: -4, maximum: 4}
            intelligence: {type: integer, minimum: -4, maximum: 4}
            charisma: {type: integer, minimum: -4, maximum: 4}
        abilities: {$ref: '#/components/schemas/CharacterSkillsInput'}
        equipment:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              maxItems: 20
              items:
                type: object
                additionalProperties: false
                required: [id, name, description]
                properties:
                  id: {type: string, maxLength: 80, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
                  name: {type: string, minLength: 1, maxLength: 120, pattern: '\S'}
                  description: {type: string, maxLength: 2000}
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
...
    Ability:
      type: object
      required: [id, name, trigger, effectText]
      properties:
        id: {type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
        name: {type: string}
        description: {type: string}
        effectText: {type: string}
        trigger: {type: string}
        uses:
          type: [object, 'null']
          properties:
            scope: {type: string, enum: [turn, round, battle, location, campaign]}
            max: {type: integer, minimum: 1}
        automationMode: {type: string, const: manual}
        effects: {type: array, maxItems: 0, items: {type: object}}
      description: Описательный навык старой анкеты v1 или повествовательная особенность v2. Формальные эффекты отсутствуют.
    Item:
      type: object
      required: [id, name]
      properties:
        id: {type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
        name: {type: string}
        description: {type: string}
        effectText: {type: string}
        charges:
          type: [object, 'null']
          properties:
            scope: {type: string, enum: [turn, round, battle, location, campaign]}
            max: {type: integer, minimum: 1}
    Validation:
      type: object
      required: [valid, issues]
      properties:
        valid: {type: boolean}
        issues:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path: {type: string}
              code: {type: string}
              message: {type: string}
        skills: {$ref: '#/components/schemas/SkillValidation'}
        derived:
          type: object
          properties:
            maxHp: {type: integer}
            baseAc: {type: integer}
            pointsSpent: {type: integer}
    CharacterSummary:
      type: object
      required: [id, displayName, raceId, classId, maxHp, baseAc, createdAt]
      properties:
        id: {type: string, format: uuid}
        displayName: {type: string}
        raceId: {type: string}
        classId: {type: string}
        portraitUrl:
          type: [string, 'null']
          format: uri-reference
          description: Путь /assets/<uuid> проверенного портрета на API. Без портрета поле отсутствует; клиент использует заглушку.
        maxHp: {type: integer}
        baseAc: {type: integer}
        createdAt: {type: string, format: date-time}
    Character:
      allOf:
        - $ref: '#/components/schemas/CharacterSummary'
        - type: object
          properties:
            pronouns: {type: string}
            roleLabel: {type: string}
            story: {type: string}
            motivation: {type: string}
            appearance: {type: string}
            personality:
              type: array
              items: {type: string}
            attributes:
              type: object
              additionalProperties: {type: integer}
            abilities:
              type: array
              items:
                anyOf:
                  - {$ref: '#/components/schemas/NormalizedAbility'}
                  - {$ref: '#/components/schemas/Ability'}
            equipment:
              type: array
              items: {$ref: '#/components/schemas/Item'}
            rulesetId: {type: string}
    ModifierStat:
      type: string
      enum: [strength, dexterity, wisdom, intelligence, charisma]
    ProfileId:
      type: string
      enum:
        - damage-d8-twice
        - damage-2d6-once
        - damage-2d8-once
        - healing-d6-twice
        - healing-d8-twice
        - healing-2d6-once
        - healing-2d8-once
    BasicActionInput:
      type: object
      additionalProperties: false
      required: [name, description, modifierStat]
      properties:
        name: {type: string, minLength: 1, maxLength: 120, pattern: '\S'}
        description: {type: string, maxLength: 2000}
        modifierStat: {$ref: '#/components/schemas/ModifierStat'}
    CustomAbilityInput:
      type: object
      additionalProperties: false
      required: [id, name, description, profileId, modifierStat]
      properties:
        id: {type: string, maxLength: 80, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
        name: {type: string, minLength: 1, maxLength: 120, pattern: '\S'}
        description: {type: string, maxLength: 2000}
        profileId: {$ref: '#/components/schemas/ProfileId'}
        modifierStat: {$ref: '#/components/schemas/ModifierStat'}
        iconAssetId:
          type: [string, 'null']
          format: uuid
          description: Необязательная иконка; использовать после подключения хранилища.
    NarrativeAbilityInput:
      type: object
      additionalProperties: false
      required: [id, name, description]
      properties:
        id: {type: string, maxLength: 80, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
        name: {type: string, minLength: 1, maxLength: 120, pattern: '\S'}
        description: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
      description: Повествовательная особенность без формального игрового усиления.
    CharacterSkillsInput:
      type: object
      additionalProperties: false
      required: [basicAction, items, narrativeItems]
      properties:
        basicAction: {$ref: '#/components/schemas/BasicActionInput'}
        items:
          type: array
          maxItems: 3
          items: {$ref: '#/components/schemas/CustomAbilityInput'}
        narrativeItems:
          type: array
          maxItems: 3
          items: {$ref: '#/components/schemas/NarrativeAbilityInput'}
      x-budget-points: 6
      x-unique-fields: [items.id, items.profileId, narrativeItems.id]
      description: |
        Полный раздел abilities для validate/complete анкеты character-creation-v2 или character-creation-v3.
        Сервер проверяет уникальность ID во всём разделе, включая reserved basic-attack,
        уникальность профилей, бюджет и существование выбранной характеристики.
        Частичный черновик можно сохранять без соблюдения required полей;
        эта полная схема применяется при validate/complete.
...
    SkillValidation:
      type: object
      additionalProperties: false
      required: [valid, pointsSpent, pointsBudget, issues]
      properties:
        valid: {type: boolean}
        pointsSpent: {type: integer, minimum: 0}
        pointsBudget: {type: integer, const: 6}
        issues:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [path, code, message]
            properties:
              path: {type: string}
              code:
                type: string
                enum: [required, invalid, unknown_profile, unknown_stat, duplicate_id, duplicate_profile, too_many, too_long, budget, asset_unavailable]
              message: {type: string}
    NormalizedAbility:
      type: object
      additionalProperties: false
      required: [id, name, description, profileId, modifierStat, effectText, automationMode, trigger, effects, uses]
      properties:
        id: {type: string}
        name: {type: string}
        description: {type: string}
        profileId:
          anyOf:
            - {$ref: '#/components/schemas/ProfileId'}
            - {type: string, const: basic-attack}
        modifierStat: {$ref: '#/components/schemas/ModifierStat'}
        effectText: {type: string, readOnly: true}
        automationMode: {type: string, const: automatic, readOnly: true}
        trigger: {type: string, const: action, readOnly: true}
        effects:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: object
            additionalProperties: false
            required: [rulesetId, profileId, type, target, dice, modifierStat, modifier]
            properties:
              rulesetId: {type: string, const: character-abilities-v1}
              profileId:
                anyOf:
                  - {$ref: '#/components/schemas/ProfileId'}
                  - {type: string, const: basic-attack}
              modifier: {type: integer}
              attackBonus: {type: integer, description: Присутствует только у урона, уже включает выбранный модификатор}
              type: {type: string, enum: [damage, healing]}
              target: {type: string, enum: [one-enemy, one-friendly]}
              dice:
                type: object
                additionalProperties: false
                required: [count, sides]
                properties:
                  count: {type: integer, minimum: 1, maximum: 2}
                  sides: {type: integer, enum: [6, 8]}
              modifierStat: {$ref: '#/components/schemas/ModifierStat'}
        uses:
          type: [object, 'null']
          additionalProperties: false
          required: [scope, max]
          properties:
            scope: {type: string, const: battle}
            max: {type: integer, minimum: 1, maximum: 2}
        iconAssetId: {type: [string, 'null'], format: uuid}
        iconUrl: {type: string, format: uri-reference}
      description: |
        Результат серверной нормализации, не вход для PATCH.
        Автоматическая механика определена структурированными данными;
        её подключение к игровому runtime — последующий этап.
...
  responses:
...
    DraftsRequireAuthentication:
      description: Черновики недоступны до подключения авторизации пользователей.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code: {type: string, const: drafts_require_authentication}
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    NotFound:
      description: Объект не найден или токен не подходит
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unavailable:
      description: Зависимость пока не настроена
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}